requirements.txt
PyMuPDF
pytesseract
faiss-cpu
sentence-transformers
gradio
google-generativeai
langchain-text-splitters
numpy
Pillow